+++ gpc-api-facade/src/main/java/uk/nhs/adaptors/pss/gpc/config/filter/package-info.java
@@ -1,0 +1,4 @@
+@NonNullApi
+package uk.nhs.adaptors.pss.gpc.config.filter;
+
+import org.springframework.lang.NonNullApi;